pyproject.toml
[project]
name = "bespokeslab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bespokelabs-curator>=0.1.26",
    "dspy>=2.6.27",
    "gepa>=0.0.22",
    "pydantic>=2.12.4",
]

[dependency-groups]
dev = [
    "ruff>=0.14.5",
]
